Fix coverity issues (#14480)
* Fix coverity issues
382921
382924
382927
382928
382932
382933
382950
382990
383123
382952
382906
382908
382912
382914
382917
382918
382919

* 381508 Unchecked return value

* 382965 Dereference after null check

+++ database/engine/metric.c
@@ -412,10 +412,10 @@ void mrg_metric_expand_retention(MRG *mrg __maybe_unused, METRIC *metric, time_t
         metric->latest_time_s_clean = last_time_s;
 
         if(likely(update_every_s))
-            metric->latest_update_every_s = update_every_s;
+            metric->latest_update_every_s = (uint32_t) update_every_s;
     }
     else if(unlikely(!metric->latest_update_every_s && update_every_s))
-        metric->latest_update_every_s = update_every_s;
+        metric->latest_update_every_s = (uint32_t) update_every_s;
 
     metric_has_retention_unsafe(mrg, metric);
     netdata_spinlock_unlock(&metric->spinlock);
@@ -578,7 +578,7 @@ bool mrg_metric_set_update_every(MRG *mrg __maybe_unused, METRIC *metric, time_t
         return false;
 
     netdata_spinlock_lock(&metric->spinlock);
-    metric->latest_update_every_s = update_every_s;
+    metric->latest_update_every_s = (uint32_t) update_every_s;
     netdata_spinlock_unlock(&metric->spinlock);
 
     return true;
@@ -590,7 +590,7 @@ bool mrg_metric_set_update_every_s_if_zero(MRG *mrg __maybe_unused, METRIC *metr
 
     netdata_spinlock_lock(&metric->spinlock);
     if(!metric->latest_update_every_s)
-        metric->latest_update_every_s = update_every_s;
+        metric->latest_update_every_s = (uint32_t) update_every_s;
     netdata_spinlock_unlock(&metric->spinlock);
 
     return true;
